requirements.txt
python-telegram-bot==21.6
discogs-client==2.3.0
anthropic>=0.40.0
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
